% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.r
\name{bottleNeck_traj}
\alias{bottleNeck_traj}
\title{Bottleneck trajectory}
\usage{
bottleNeck_traj(t, ne.max = 100, ne.min = 10, bstart = 2, bend = 1,
  ...)
}
\description{
Piecewise constant trajectory with two levels where the lower level represents a population bottleneck.
}
