% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_functions.r
\name{coaltimeSim}
\alias{coaltimeSim}
\title{Simulate from inhomogeneous, heterochronous coalescent.}
\usage{
coaltimeSim(samp_times, n_sampled, traj, lower_bound, ...)
}
\arguments{
\item{samp_times}{numeric vector of sampling times.}

\item{n_sampled}{numeric vector of samples taken per sampling time.}

\item{traj}{function that returns effective population size at time t.}

\item{lower_bound}{numeric lower limit on \code{traj} function on its support.}

\item{...}{additional arguments to be passed to \code{traj} function.}
}
\value{
A list containing vectors of coalescent times \code{coal_times}, 
  intercoalescent times \code{intercoal_times}, and number of active lineages
  \code{lineages}, as well as passing along \code{samp_times} and
  \code{n_sampled}.
}
\description{
For simulating coalescent times using the method of thinning. 
This function is based on code written by Michael Karcher for the \code{phylodyn} package.
}
\examples{
coaltimeSim(0:2, 3:1, unif_traj, lower_bound=10, level=10)
}
