% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_functions.r
\name{coaltimeSimGP}
\alias{coaltimeSimGP}
\title{Simulate from inhomogeneous, heterochronous coalescent driven by Gaussian process.}
\usage{
coaltimeSimGP(samp_times, n_sampled, trajvec, tvec, lower_bound)
}
\arguments{
\item{samp_times}{numeric vector of sampling times.}

\item{n_sampled}{numeric vector of samples taken per sampling time.}

\item{trajvec}{vector of population sizes (should be dense)}

\item{tvec}{grid boundaries for trajvec. Length is length(trajvec)+1}

\item{lower_bound}{numeric lower limit on \code{traj} function on its support.}
}
\value{
A list containing vectors of coalescent times \code{coal_times}, 
  intercoalescent times \code{intercoal_times}, and number of active lineages
  \code{lineages}, as well as passing along \code{samp_times} and
  \code{n_sampled}.
}
\description{
For simulating coalescent times using the method of thinning. 
This function is based on code written by Michael Karcher for the \code{phylodyn} package.
}
