% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_theta.r
\name{extract_theta}
\alias{extract_theta}
\title{Extract latent process parameters and associated Bayesian credible intervals}
\usage{
extract_theta(mfit, obstype = "normal", alpha = 0.05)
}
\arguments{
\item{mfit}{An object containing posterior draws from a \code{bnps} model or \code{rstan::stan} model fit.  The object can be of class 'stanfit', 'array', 'matrix', or 'data.frame'.}

\item{obstype}{Distribution name of observations.  This controls the back-transformation of the process parameters.  Possible obstypes are 'normal', 'poisson', or 'binomial'.}

\item{alpha}{Controls level for (1-alpha) Bayesian credible intervals.}
}
\value{
Returns the posterior median and posterior (1-alpha) quantiles of theta.
}
\description{
Extract theta (latent process) parameters from an object with posterior draws and associated Bayesian credible intervals.
}
\seealso{
rstan, rstan::as.array, rstan::as.matrix, rstan::as.data.frame
}

