% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_init.r
\name{get_init}
\alias{get_init}
\title{Generate function for initial parameter values}
\usage{
get_init(prior = "horseshoe", likelihood = "normal", order = 1)
}
\arguments{
\item{prior}{A character string specifying which prior to use on order-\emph{k} differences. Choices are "horseshoe", "laplace", and "normal".}

\item{likelihood}{A character string specifying the distribution of the observations. Choices are "normal", "poisson", and "binomial".}

\item{order}{Numeric value specifying order of differencing (1, 2, or 3).}
}
\value{
An object of class \code{function} for generating initial parameter seeds for passage to \code{stan} via \code{bnps}. Parameters are appropriate to a model with specified prior, likelihood, and order.
}
\description{
Creates a function which generates initial parameter seeds for passing to \code{stan} internally via \code{bnps}.
}
\seealso{
\code{\link{bnps}}, \code{\link[rstan]{stan}}, \code{\link{get_model}}
}

