% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model.r
\name{get_model}
\alias{get_model}
\title{Generate model code for passage to stan}
\usage{
get_model(prior = "horseshoe", likelihood = "normal", order = 1,
  zeta = 0.01)
}
\arguments{
\item{prior}{A character string specifying which prior to use on order-k differences. Choices are "horseshoe", "laplace", and "normal".}

\item{likelihood}{A character string specifying the probability distribution of the observation variable. Current choices are "normal", "poisson", and "binomial".}

\item{order}{Numeric value specifying order of differencing (1, 2, or 3).}

\item{zeta}{The hyperparameter for the global smoothing parameter gamma.  This is the scale parameter of a half-Cauchy distribution.  Smaller values will result in more smoothing depending on the prior specification and the strength of the data. Values must be > 0.}
}
\value{
A function for generating initial parameter seeds for passage to stan. Parameters are appropriate to a model with specified prior, likelihood, and order.
}
\description{
Generates a text string of code describing the model of interest for passage to the stan function in the rstan package.
}
\seealso{
getInit, rstan::stan
}

