% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_coalescent_data.r
\name{make_coalescent_data}
\alias{make_coalescent_data}
\title{Make data list for coalescent data}
\usage{
make_coalescent_data(samp_times, n_sampled, coal_times, grid)
}
\arguments{
\item{samp_times}{numeric vector of sampling times.}

\item{n_sampled}{numeric vector of samples taken per sampling time.}

\item{coal_times}{vector of coalescent times}

\item{grid}{vector of grid boundaries (length is number of grid cells plus 1)}
}
\value{
A list containing 
\itemize{
\item{\code{J}}{ number of grid cells}
\item{\code{N}}{ number of subgrid cells}
\item{\code{y}}{ vector  of length \code{N} of binary indicators for whether a subgrid cell ends in a coalescent time or not}
\item{\code{gridrep}}{ vector  of length \code{J} containing number of subintervals in each grid cell}
\item{\code{Aik}} { vector of length \code{N} containing the binomial coefficient or 'coalescent factor' for each subgrid cell. This is equal to choose(x.k, 2), where x.k is the number of active lineages in subgrid cell k. }
\item{\code{coalind}} { vector of length \code{N} containing indicator for the coalescent event associated with each subgrid cell}
\item{\code{ncoal}} { number of coalescent times, which is equal to the number of samples minus 1}
\item{\code{ncoalv}} { vector of length \code{J} containing the number of coalescent events in each grid cell }
\item{\code{dalpha}}  { vector of length \code{N} containing the width of each subgrid cell }
\item{\code{rep.idx}} { matrix of \code{J} by 2 where the first column contains for the index for the subgrid cell at which each grid cell starts and the second column contains the index for the subgrid cell at which each grid cell ends.  }
\item{\code{log_mu}} { scalar value containing the natural log of the MLE for a constant population size estimated from the coalescent data}
}
}
\description{
Generates a list containing the data elements necessary to fit a SPMRF model to coalescent data, where effective population sizes are estimated on a uniform grid.
This function is based on code written by Michael Karcher for the \code{phylodyn} package and expanded for use with \code{spmrf}.
}
