% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_functions.r
\name{makeGrid}
\alias{makeGrid}
\title{Make grid for coalescent data}
\usage{
makeGrid(coal_times, samp_times, Ngrid)
}
\description{
Makes grid for effective population size estimation from coalescent data.
}
