% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.r
\name{piecewiseExp_traj}
\alias{piecewiseExp_traj}
\title{Piecewise-exponential trajectory}
\usage{
piecewiseExp_traj(t, tbreaks, lnvals, reverse = FALSE, tmx = NULL)
}
\description{
Piecewise-exponential trajectory
}
