% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trace.r
\name{plot_trace}
\alias{plot_trace}
\title{Plot traces of posterior draws of a parameter}
\usage{
plot_trace(postob, vname, pscale = "original", stack = FALSE,
  colset = "color")
}
\arguments{
\item{postob}{An array containing posterior draws from a \code{bnps} model fit or \code{rstan::stan} model fit.  Model fit objects of type 'stanfit' can be coerced into arrays using either the 'extract' or 'as.array' functions from the \code{rstan} package.}

\item{vname}{Name of variable to plot.  Name must correspond to a dimname in the array.}

\item{pscale}{Plotting scale. Options are 'original', 'log', or 'inv' (inverse).}

\item{stack}{Indicator to plot chains separately and stack the plots.}

\item{colset}{Color scheme for plotting traces by chain.  Schemes 'color' and 'gray' will produce separate colors or gray shades for each chain, and scheme 'black' will make all chain traces black.}
}
\value{
Returns a plot object
}
\description{
Plots the posterior draws of a specified parameter by MCMC iteration with options for visual differentiation of separate chains.
}

