% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trend.r
\name{plot_trend}
\alias{plot_trend}
\title{Plot estimated trend and credible intervals}
\usage{
plot_trend(theta = NULL, mfit = NULL, obstype = "normal", alpha = 0.05,
  obsvec = NULL, timelab = NULL, colset = "blue", trend.col = NULL,
  bci.col = NULL, ...)
}
\arguments{
\item{theta}{A list generated by \code{extractTheta} function.  If \code{theta} is specified, do not specify a value for \code{mfit}.}

\item{mfit}{An object containing posterior draws from a \code{bsmfit} or \code{stan} model fit.  The object can be of class 'stanfit', 'array', 'matrix', or 'data.frame'.  Do not specify \code{theta} if \code{mfit} is specified.}

\item{obstype}{Name of probability distribution of observations.  This controls the back-transformation of the process parameters.  Possible obstypes are 'normal', 'poisson', or 'binomial'. Assumed link functions are the log for 'poisson' and logit for 'binomial'.}

\item{alpha}{Controls level for (1-alpha) Bayesian credible intervals.}

\item{obsvec}{Vector of observations. Assumed to be ordered by time with one observation per time point. If obsvec is NULL, then data are not plotted.}

\item{timelab}{Label for time variable.  Can be a numeric or character vector.  Must be the same length as observation vector.  If not specified, then a sequence from 1 to the total number of observations with unit time step is used.}

\item{colset}{Color set for trend and credible intervals.  These are presets chosen with common color combinations.  Choices are 'blue', 'purple', 'red', 'green', and 'gray'. This parameter is overridden If trend.col and bci.col are specified.}

\item{trend.col}{Color of the trend line.  Any standard \code{R} plotting colors are allowed. Use this along with bci.col if color choices other than the presets provided by colset are desired.}

\item{bci.col}{Color of the credible interval band. Any standard \code{R} colors are allowed.}

\item{...}{Additional parameters passed to the plot function.}
}
\value{
Returns a plot object.
}
\description{
Plot posterior median of trend parameters and associated credible intervals.
}
\details{
The input can be either in the form of a list generated by the \code{extractTheta} function (passed in as the \code{theta} variable) or as a raw model fit object from \code{bnps} or \code{rstan} (passed in as \code{mfit}).  Note that using \code{mfit} is more computationally costly than using \code{theta} because the \code{extract_theta} function is called when \code{mfit} is specified.
}

