% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_functions.r
\name{skyLine}
\alias{skyLine}
\title{Classic Skyline Estimates}
\usage{
skyLine(phy)
}
\arguments{
\item{phy}{a list containing a numeric vector of coalescent times (\code{coal_times}), a numeric vector of sampling times (\code{samp_times}, and a integer vector of number of samples taken at each sampling time (\code{n_sampled}).}
}
\value{
A list containing a vector of indices \code{k}, a vector of midpoints between coalescent times \code{mid_ctime}, and a vector of skyline population estimates \code{theta}.
}
\description{
Classic Skyline estimates for isochronous or heterochronous data.
}
