% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varRef.r
\name{varRef}
\alias{varRef}
\title{Calculate reference marginal variance of field parameters}
\usage{
varRef(nn, kap, omg2, order = 1)
}
\description{
This is an internal function used by the \code{set_zeta} function
}
\keyword{internal}

