% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnps.r
\name{bnps}
\alias{bnps}
\title{Fit Bayesian nonparametric adaptive smoothing model}
\usage{
bnps(initialize = FALSE, likelihood = "normal", prior = "normal",
  order = 1, zeta = 0.01, file, model_name = "anon_model",
  model_code = "", fit = NA, data = list(), pars = NA, chains = 4,
  iter = 2000, warmup = floor(iter/2), thin = 1, init = "random", ...)
}
\description{
Fit Bayesian nonparametric adaptive smoothing model
}

