% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_init.r
\name{get_init}
\alias{get_init}
\title{Generate function for initial parameter values}
\usage{
get_init(prior = "horseshoe", likelihood = "normal", order = 1)
}
\arguments{
\item{prior}{A character string specifying which prior to use on order-k differences. Choices are "horseshoe", "laplace", and "normal".}

\item{likelihood}{A character string specifying the distribution of the observations. Choices are "normal", "poisson", and "binomial".}

\item{order}{Numeric value specifying order of differencing (1, 2, or 3).}
}
\value{
A function for generating initial parameter seeds for passage to stan. Parameters are appropriate to a model with specified prior, likelihood, and order.
}
\description{
Creates a function for passing to stan which generates initial parameter seeds.
}
\seealso{
getModel, rstan::stan
}

