% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model.r
\name{get_model}
\alias{get_model}
\title{Generate model code for passage to stan}
\usage{
get_model(prior = "horseshoe", likelihood = "normal", order = 1,
  zeta = 0.01, save.loglik = FALSE)
}
\arguments{
\item{prior}{A character string specifying which prior to use on order-k differences. Choices are "horseshoe", "laplace", and "normal". Note that "laplace" priors are currently not available for coalescent likelihoods.}

\item{likelihood}{A character string specifying the probability distribution of the observation variable. Current choices are "normal", "poisson", "binomial", and "coalescent".}

\item{order}{Numeric value specifying order of differencing (1, 2, or 3). Note that order 3 is currently not available for coalescent likelihoods.}

\item{zeta}{The hyperparameter for the global smoothing parameter gamma.  This is the scale parameter of a half-Cauchy distribution.  Smaller values will result in more smoothing depending on the prior specification and the strength of the data. Values must be > 0.}

\item{save.loglik}{Logical flag for whether to calculate the individual components of the log-likelihood for use in calculating WAIC or LOOIC with the \code{loo} package. Default is FALSE.}
}
\value{
A character string of code describing the model of interest for passage to the function \code{stan} in the \code{rstan} package.
}
\description{
Generates a text string of code describing the model of interest for passage to the \code{stan} function in the \pkg{rstan} package.  This function is called internally in the \code{spmrf} function.
}
\details{
This function can be used to generate a text string containing code in the \code{stan} model syntax.  This function is called by the \code{spmrf} function internally, so it is not necessary to use \code{get_model} external to \code{spmrf}.
}
\seealso{
\code{\link{spmrf}},  \code{\link[rstan]{stan}}, \code{\link{get_init}}
}
