% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model.r
\name{get_model}
\alias{get_model}
\title{Generate model code for passage to stan}
\usage{
get_model(prior = "horseshoe", likelihood = "normal", order = 1,
  zeta = 0.01)
}
\arguments{
\item{prior}{A character string specifying which prior to use on order-k differences. Choices are "horseshoe", "laplace", and "normal".}

\item{likelihood}{A character string specifying the probability distribution of the observation variable. Current choices are "normal", "poisson", and "binomial".}

\item{order}{Numeric value specifying order of differencing (1, 2, or 3).}

\item{zeta}{The hyperparameter for the global smoothing parameter gamma.  This is the scale parameter of a half-Cauchy distribution.  Smaller values will result in more smoothing depending on the prior specification and the strength of the data. Values must be > 0.}
}
\value{
A character string of code describing the model of interest for passage to the function \code{stan} in the \code{rstan} package.
}
\description{
Generates a text string of code describing the model of interest for passage to the \code{stan} function in the \pkg{rstan} package.  This function is called internally in the \code{bnps} function.
}
\details{
This function can be used to generate a text string containing code in the \code{stan} model syntax.  This function is called by the \code{bnps} function internally, so it is not necessary to use \code{get_model} external to \code{bnps}.
}
\seealso{
\code{\link{bnps}},  \code{\link[rstan]{stan}}, \code{\link{get_init}}
}

