% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trace.r
\name{plot_trace}
\alias{plot_trace}
\title{Plot traces of posterior draws of a parameter}
\usage{
plot_trace(postob, vname, pscale = "original", stack = FALSE,
  colset = "color")
}
\arguments{
\item{postob}{An array containing posterior draws from a \code{spmrf} model fit or \code{stan} model fit.  Model fit objects of type \code{stanfit} can be coerced into arrays using either the \code{extract} or \code{as.array.stanfit} functions from the \pkg{rstan} package.}

\item{vname}{Name of variable to plot.  Name must correspond to a dimname in the array.}

\item{pscale}{Plotting scale. Options are 'original', 'log', or 'inv' (inverse).}

\item{stack}{Indicator to plot chains separately and stack the plots.}

\item{colset}{Color scheme for plotting traces by chain.  Schemes 'color' and 'gray' will produce separate colors or gray shades for each chain, and scheme 'black' will make all chain traces black.}
}
\value{
Returns a \code{plot} object
}
\description{
Plots the posterior draws of a specified parameter by MCMC iteration with options for visual differentiation of separate chains.
}
\seealso{
\code{\link[rstan]{stanfit}}, \code{\link[rstan]{extract}}, \code{\link[rstan]{as.array.stanfit}}, \code{\link{spmrf}}
}

