% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trend.r
\name{plot_trend}
\alias{plot_trend}
\title{Plot estimated trend and credible intervals}
\usage{
plot_trend(theta = NULL, mfit = NULL, obstype = "normal",
  alpha = 0.05, obsvar = NULL, uxvar = NULL, xvar = NULL,
  colset = "blue", trend.col = NULL, trend.lwd = 3, bci.col = NULL,
  pt.col = "gray60", pt.pch = 1, pt.cex = 1, ylab, xlab, main, ylim,
  xlim, ...)
}
\arguments{
\item{theta}{A list generated by \code{extract_theta} function.  If \code{theta} is specified, do not specify a value for \code{mfit}.}

\item{mfit}{An object containing posterior draws from a \code{spmrf} or \code{stan} model fit.  The object can be of class \code{stanfit}, \code{array}, \code{matrix}, or \code{data.frame}.  Do not specify \code{theta} if \code{mfit} is specified.}

\item{obstype}{Name of probability distribution of observations.  This controls the back-transformation of the process parameters.  Possible obstypes are 'normal', 'poisson', 'binomial', or 'coalescent'. Assumed link functions are the log for 'poisson' and 'coalescent', and logit for 'binomial'.}

\item{alpha}{Controls level for 100*(1-\code{alpha})\% Bayesian credible intervals.}

\item{obsvar}{Vector of observations. Assumed to be ordered by time with one observation per time point. If \code{obsvar} is NULL, then data are not plotted. Note if plotting observations you must also specify a matching \code{xvar} of the same length.}

\item{uxvar}{vector of unique values of the grid locations of the trend parameters. This can be a numeric or integer vector. Must be the same length as \code{theta} and must be sorted to match order of locations of \code{theta} elements. If not specified and no \code{xvar} is specified, then a sequence from 1 to the length of \code{theta} with unit spacing will be used.  If \code{xvar} and \code{obsvar} are specified, then \code{uxvar} does not need to be specified.}

\item{xvar}{x-variable for plotting observations \code{obsvar}. Should be a numeric integer vector.  Must be the same length as observation vector (\code{obsvar}). Must be specified if \code{obsvar} is specified.}

\item{colset}{Color set for trend and credible intervals.  These are presets chosen with common color combinations.  Choices are 'blue', 'purple', 'red', 'green', and 'gray'. This parameter is overridden If \code{trend.col} and \code{bci.col} are specified.}

\item{trend.col}{Color of the trend line.  Any standard \code{R} plotting colors are allowed. Use this along with \code{bci.col} if color choices other than the presets provided by colset are desired.}

\item{bci.col}{Color of the credible interval band. Any standard \code{R} colors are allowed.}

\item{pt.col}{Color of data points if plotted with \code{obsvec} specified.}

\item{pt.pch}{Type of points to use if data plotted with \code{obsvec} specified (see \code{points} for details.}

\item{pt.cex}{Size of points to use if data plotted with \code{obsvec} specified (see \code{points} for details.}

\item{ylab}{Character string describing label for y-axis.  Default is "y" if unspecified.}

\item{xlab}{Character string describing label for x-axis.  Default is "x" if unspecified.}

\item{main}{Character string describing label for main title.  Default is "" empty string.}

\item{ylim}{Numeric vector with two elements describing limits of y-variable. Default is \code{range} of combined vector of \code{obsvec} and extreme values of BCI's if unspecified.}

\item{xlim}{Numeric vector with two elements describing limits of x-variable. Default is \code{range} of sequence from 1 to max length of elements in \code{theta} or number of rows in \code{mfit}.}

\item{...}{Additional parameters passed to the \code{plot} function.}
}
\value{
Returns a \code{plot} object.
}
\description{
Plot posterior median of trend parameters and associated credible intervals.
}
\details{
The input can be either in the form of a list generated by the \code{extract_theta} function (passed in as the \code{theta} variable) or as a raw model fit object from \code{spmrf} or \code{rstan} (passed in as \code{mfit}).  Note that using \code{mfit} is more computationally costly than using \code{theta} because the \code{extract_theta} function is called when \code{mfit} is specified.
}
\seealso{
\code{\link[graphics]{plot}}, \code{\link[graphics]{points}}, \code{\link{extract_theta}}, \code{\link{spmrf}}, \code{\link[rstan]{stanfit}}, \code{\link[base]{array}}, \code{\link[base]{matrix}}, \code{\link[base]{data.frame}}
}
