% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spmrf.r
\name{spmrf}
\alias{spmrf}
\title{Fit Bayesian nonparametric adaptive smoothing model}
\usage{
spmrf(prior = "horseshoe", likelihood = "normal", order = 1, zeta = 0.01,
  fit = NA, data, pars = NA, chains = 4, iter = 2000,
  warmup = floor(iter/2), thin = 1, ...)
}
\arguments{
\item{prior}{A character string specifying which prior to use on order-\emph{k} differences. Choices are "horseshoe", "laplace", and "normal".}

\item{likelihood}{A character string specifying the probability distribution of the observation variable. Current choices are "normal", "poisson", and "binomial".}

\item{order}{Numeric value specifying order of differencing (1, 2, or 3).}

\item{zeta}{The hyperparameter for the global smoothing parameter gamma.  This is the scale parameter of a half-Cauchy distribution.  Smaller values will result in more smoothing depending on the prior specification and the strength of the data. Values must be > 0.}

\item{fit}{An instance of S4 class \code{stanfit} derived from a previous fit; defaults to NA. If \code{fit} is not NA, the compiled model associated with the fitted result is re-used; thus the time that would otherwise be spent recompiling the C++ code for the model can be saved.}

\item{data}{A named list providing the data for the model. See details below.}

\item{pars}{A vector of character strings specifying parameters of interest; defaults to NA indicating all parameters in the model. If \code{include} = TRUE, only samples for parameters given in \code{pars} are stored in the fitted results. Conversely, if \code{include} = FALSE, samples for all parameters except those given in \code{pars} are stored in the fitted results.}

\item{chains}{A positive integer specifying number of chains; defaults to 4.}

\item{iter}{A positive integer specifying how many iterations for each chain (including warmup). The default is 2000.}

\item{warmup}{A positive integer specifying number of warmup (aka burnin) iterations. This also specifies the number of iterations used for stepsize adaptation, so warmup samples should not be used for inference. The number of warmup should not be larger than \code{iter} and the default is \code{iter}/2.}

\item{thin}{A positive integer specifying the period for saving sample; defaults to 1.}

\item{...}{Additional arguments passed to \code{rstan::stan}.  Do not include \code{file}, \code{model_code}, or \code{init} in this list.}
}
\value{
An object of class \code{stanfit}.  See \code{stanfit} and \code{stan} for more details.
}
\description{
Fit Bayesian nonparametric adaptive temporal smoothing models using Hamiltonian Monte Carlo with the \code{stan} engine from the \pkg{rstan} package.
}
\details{
This function first internally creates \code{stan} model code and a function to generate initial parameter seeds using the information specified in the \code{prior}, \code{likelihood}, \code{order}, and \code{zeta} arguments.  It then passes these to the \code{stan} function.  The \code{spmrf} function will take additional arguments passed to \code{stan} \emph{except for} the arguments \code{file}, \code{model_code}, and \code{init}.  These arguments are not accepted because they may conflict with model forms expected by \code{spmrf}.  See the \code{stan} function documentation for more details.

\code{stan} does all of the work of fitting a Stan model and returning the results as an instance of \code{stanfit}. First, it translates the Stan model to C++ code. Second, the C++ code is compiled into a binary shared object, which is loaded into the current \code{R} session (an object of S4 class \code{stanmodel} is created). Finally, samples are drawn and wrapped in an object of S4 class \code{stanfit}, which provides functions such as \code{print}, \code{summary}, and \code{plot} to inspect and retrieve the results of the fitted model.

\code{spmrf} can also be used to sample again from a fitted model under different settings (e.g., different \code{iter}) by providing argument \code{fit}. In this case, the compiled C++ code for the model is reused.
}
\references{
Faulkner, J. R., and V. N. Minin. 2015. Bayesian trend filtering: adaptive temporal smoothing with shrinkage priors. \emph{arXiv preprint}.
}
\seealso{
\code{\link[rstan]{stan}}, \code{\link[rstan]{stanfit}}, \code{\link{get_model}}, and \code{\link{get_init}}
}

